1.0.4 - Do not remove tempfile if it has not been generated

+++ aws/__init__.py
@@ -312,7 +312,7 @@ class PipLayers(core.Construct):
                 with open(layer_unpack_dir / "md5sum", "w") as f:
                     f.write(req_md5)
 
-                if os.path.exists(tempname):
+                if devpackage and os.path.exists(tempname):
                     os.remove(tempname)
             else:
                 print(f"Using cached layer image for {layer_id}.")
